tercoReader/requirements.txt
streamlit
scikit-learn
transformers
rank_bm25
psycopg2-binary
torch
numpy
sqlalchemy
llama-index-llms-groq